Start playing with a provider class.
This means not just defining what a provider does, but also the
underlying class which contains some methods to implement that. It would
appear it makes most sense to follow the existing puppet model, which is
to prefetch all relevant data into a property array, and then flush it
later once we've updated that array.