Check that existing ldap data has expected class.
It's bad if we're expecting something to be a group, and what we
actually get is an inetorgperson, or something like that. So, raise an
error if this happens.

I suspect it would be /very bad/ if we deleted part of an LDAP tree just
like that because it didn't look how we expected. Can you say 'sudden
and arbitary data loss'?