Add properties avaiable for inetorgperson
So the new plan: no attempt at clever resource type definitions. We just
have an ldapres that can feature all the attributes that exist, and if
you set an attribute that isn't valid for the objectClass then the ldap
server rejects it (you actually get a useful error message).

This is because puppet doesn't appear to have any useful way for me to
make a provider apply across a class of resource types. Perhaps there is
and I haven't seen it. Also because I can't find a good way to make
custom types inherit nicely (again, I didn't look very hard).