Kill array-munger for types.
This was converting all non-arrays into a single-element array
containing the data item, so that everything entering ldap was an array
of values. However, it turns out that enabling :array_matching does that
by default.

And :array_matching is necessary for being able to specify multiple
objectClass fields, and probably for other multi-value fields.