Add method for deleting unconfigured properties.
That is; if we've changed something, then we'll go through all
properties and update the ldap resource to match them; and delete all
other properties that the ldap resource has right now. That means that
if you delete a few properties and change a few others, the deleted
properties evaporate.

Note that the one massive bug in this is that puppet has to decide that
the resource is to be updated when this happens. That means that if you
delete a property in puppet, it won't be deleted from LDAP until some
other modification to the resource is made. Which is bad.