Allow users to shoot themeslves in the foot.
That is, when the object class of something changes, happily modify the
resource without any complaints. This potentially means that somethings
objectclass changes and data becomes lost.

Ultimately this was going to hold up legitimate changes to objectclass,
so had to go. People just have to be careful and actually back up
production data (or not configure production data via puppet).