Handle some string case rubbisy
All puppet properties have to be lowercase. Attribute names come out of
ldap in a variety of cases, but internally it appears that they're
normalised to lowercase.

Currently some pain is occuring because we compare two attributes with
the same name, but different capitalisation; decide to replace one of
them, and delete the other, which is absurd and slapd chokes on it.

Fix this by always comparing in a lowercase manner.

This capitalisation goo is one of the reasons why objectclass is as
benighted as it is.